// src/types/playerImage.ts
export interface PlayerImage {
    id: number;
    image: string,
    name: string;
    matches: number;
    innings: number;
    runs: number;
    highestScore: string;
    average: number;
    strikeRate: number;
    centuries: number;
    halfCenturies: number;
    fours: number;
    sixes: number;
    wickets: number;
  }
  